refactor: api 폴더 + 공통 파일명으로 리팩터링 완료
