+++ features/profile/api/mutations.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
-import { createOnboarding, updateProfile } from './profile-api';
-import { profileKeys } from './profile-queries';
+import { createOnboarding, updateProfile } from './fetchers';
+import { profileKeys } from './queries';
 
 export const useOnboarding = () =>
   useMutation({
