+++ features/profile/index.ts
@@ -1,3 +1,3 @@
-export { useProfile } from './services/profile-queries';
-export { useOnboarding, useUpdateProfile } from './services/profile-mutations';
+export { useOnboarding, useUpdateProfile } from './api/mutations';
+export { useProfile } from './api/queries';
 export type { OnboardingRequest, Profile, UpdateProfileRequest } from './types/profile';
