+++ features/review/api/mutations.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
-import { createReview } from './review-api';
-import { reviewKeys } from './review-queries';
+import { createReview } from './fetchers';
+import { reviewKeys } from './queries';
 
 export const useCreateReview = () => {
   const queryClient = useQueryClient();
